% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cB.R
\name{cbinomMME}
\alias{cbinomMME}
\title{MME estimates for r and beta of the Continuous Binomial distribution}
\usage{
cbinomMME(data)
}
\arguments{
\item{data}{required for parameter estimation}
}
\value{
size and probability estimates of the continuous Binomial distribution by MME
}
\description{
MME estimates for r and beta of the Continuous Binomial distribution
}
\examples{
cbinomMME(data = rcbinom(n = 1000, size = 5, prob = .5))
}
