% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cB.R
\name{qq.cbinom}
\alias{qq.cbinom}
\title{Quantile-Quantile plot for continuous Binomial distribution}
\usage{
qq.cbinom(data, correction = 0.25)
}
\arguments{
\item{data}{required to check for model fitting}

\item{correction}{tail correction value}
}
\value{
a quantile-quantile plot
}
\description{
Quantile-Quantile plot for continuous Binomial distribution
}
\examples{
qq.cbinom(data = rcbinom(n = 25, size = 5, prob = .4))
}
