% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cN.R
\name{qq.cnbinom}
\alias{qq.cnbinom}
\title{Quantile-Quantile plot for Continuous Negative Binomail distribution}
\usage{
qq.cnbinom(data, correction = 0.25)
}
\arguments{
\item{data}{required for model fitting}

\item{correction}{tail correction value}
}
\value{
a quantile-quantile plot
}
\description{
Quantile-Quantile plot for Continuous Negative Binomail distribution
}
\examples{
qq.cnbinom(data = rcnbinom(100, r = 2, beta = 3))
}
